+++ lib/custom_refresh_indicator.dart
@@ -1,5 +1,6 @@
 export 'src/data/data.dart';
 export 'src/custom_refresh_indicator.dart';
+export 'src/indicator_controller.dart';
 export 'src/widgets/widgets.dart';
 export 'src/physics/physics.dart';
 export 'src/utils/utils.dart';