+++ lib/src/indicator_controller.dart
@@ -1,4 +1,7 @@
-part of 'custom_refresh_indicator.dart';
+import 'package:custom_refresh_indicator/custom_refresh_indicator.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:meta/meta.dart';
 
 class IndicatorController extends Animation<double>
     with